move accound refound view to counter
